backend/pyproject.toml
[project]
name = "qu"
version = "0.1.0"
description = "Profiler quant agent for the Linex loyalty platform"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0",
    "mcp>=1.7",
    "pandas>=2.0",
    "pydantic>=2.0",
    "firebase-admin>=6.2.0",
    "firebase-functions>=0.1.0",
    "functions-framework>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
include = ["analysis*", "cards*", "models*", "prompts*", "utils*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
